pyproject.toml
[project]
name = "gtfs-rt-archiver"
version = "0.1.0"
description = "A lightweight, resilient service for archiving GTFS-Realtime feeds to cloud storage"
license = "AGPL-3.0-only"
readme = "README.md"
authors = [
    { name = "Chris Alfano", email = "chris@jarv.us" }
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "apscheduler>=4.0.0a6",
    "gcloud-aio-storage>=9.6.1",
    "google-cloud-secret-manager>=2.26.0",
    "httpx>=0.28.0,<1.0",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
    "tenacity>=9.1.2",
]

[project.scripts]
gtfs-rt-archiver = "gtfs_rt_archiver:main"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "respx>=0.22.0",
    "ruff>=0.14.10",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
    "gcloud.*",
    "apscheduler.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
